% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{assemble_disclosure_d11}
\alias{assemble_disclosure_d11}
\title{Assemble Disclosure D1-1 Table}
\usage{
assemble_disclosure_d11(geog = "County Total")
}
\arguments{
\item{geog}{The geography (or report_level) for which the data
are to be pulled. See the factor levels for that variable for
options.}
}
\value{
A tibble of data
}
\description{
This function reads in all available CRA data and assembles
a joint tibble with the available data.
}
