% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{import_disclosure_file_d11}
\alias{import_disclosure_file_d11}
\title{Import Disclosure File D1-1}
\usage{
import_disclosure_file_d11(year)
}
\arguments{
\item{year}{A 4-digit year}
}
\value{
A tibble of disclosure data
}
\description{
This function will read in the D1-1 disclosure files
for a given calendar year.
}
