% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{import_transmittal_file}
\alias{import_transmittal_file}
\title{Import Transmittal File}
\usage{
import_transmittal_file(year)
}
\arguments{
\item{year}{The year of data to be read (can be 2- or 4-digit years)}
}
\value{
A tibble of data
}
\description{
This function imports the transmittal file for each year of CRA
data.
}
