% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{to_logical}
\alias{to_logical}
\title{Convert Character to Logical}
\usage{
to_logical(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A logical vector
}
\description{
This function converts a character vector into a
logical value, where 'Y' is TRUE, 'N' is FALSE, and
everything else is set to NA.
}
