% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{to_numeric}
\alias{to_numeric}
\title{Convert Character to Numeric}
\usage{
to_numeric(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A numeric vector
}
\description{
This function converts a character vector to a numeric
value. Blank or 'NA' values are first converted to missing
values.
}
